Factor out CountNodeVisitor class into separate files.
