+++ src/intcomp/RemoveNodesVisitor.h
@@ -20,7 +20,7 @@
 #ifndef DECOMPRESSOR_SRC_INTCOMP_REMOVENODESVISITOR_H
 #define DECOMPRESSOR_SRC_INTCOMP_REMOVENODESVISITOR_H
 
-#include "intcomp/CountNode.h"
+#include "intcomp/CountNodeVisitor.h"
 
 namespace wasm {
